EduSemestr  adding credit

+++ console/migrations/m211012_125050_edu_semestr_subject.php
@@ -14,27 +14,26 @@ class m211012_125050_edu_semestr_subject extends Migration
     {
         $this->createTable('edu_semestr_subject', [
             'id' => $this->primaryKey(),
-            'edu_semestr_id'=>$this->integer()->notNull(),
-            'subject_id'=>$this->integer()->notNull(),
-            'subject_type_id'=>$this->integer()->Null(),
-            'credit'=>$this->float()->defaultValue(0),
-            'all_ball_yuklama'=>$this->integer()->defaultValue(0),
-            'is_checked'=>$this->integer()->defaultValue(0),
-            'max_ball'=>$this->integer()->defaultValue(0),
+            'edu_semestr_id' => $this->integer()->notNull(),
+            'subject_id' => $this->integer()->notNull(),
+            'subject_type_id' => $this->integer()->Null(),
+            'credit' => $this->double()->defaultValue(0),
+            'all_ball_yuklama' => $this->integer()->defaultValue(0),
+            'is_checked' => $this->integer()->defaultValue(0),
+            'max_ball' => $this->integer()->defaultValue(0),
 
-            'order'=>$this->tinyInteger(1)->defaultValue(1),
+            'order' => $this->tinyInteger(1)->defaultValue(1),
             'status' => $this->tinyInteger(1)->defaultValue(0),
-            'created_at'=>$this->integer()->notNull(),
-            'updated_at'=>$this->integer()->notNull(),
+            'created_at' => $this->integer()->notNull(),
+            'updated_at' => $this->integer()->notNull(),
             'created_by' => $this->integer()->notNull()->defaultValue(0),
             'updated_by' => $this->integer()->notNull()->defaultValue(0),
             'is_deleted' => $this->tinyInteger()->notNull()->defaultValue(0),
         ]);
 
-        $this->addForeignKey('je_edu_semestr_subject_edu_semestr_id','edu_semestr_subject','edu_semestr_id','edu_semestr','id');
-        $this->addForeignKey('se_edu_semestr_subject_subject_id','edu_semestr_subject','subject_id','subject','id');
-        $this->addForeignKey('te_edu_semestr_subject_subject_type_id','edu_semestr_subject','subject_type_id','subject_type','id');
-
+        $this->addForeignKey('je_edu_semestr_subject_edu_semestr_id', 'edu_semestr_subject', 'edu_semestr_id', 'edu_semestr', 'id');
+        $this->addForeignKey('se_edu_semestr_subject_subject_id', 'edu_semestr_subject', 'subject_id', 'subject', 'id');
+        $this->addForeignKey('te_edu_semestr_subject_subject_type_id', 'edu_semestr_subject', 'subject_type_id', 'subject_type', 'id');
     }
 
     /**
@@ -42,9 +41,9 @@ class m211012_125050_edu_semestr_subject extends Migration
      */
     public function safeDown()
     {
-        $this->dropForeignKey('je_edu_semestr_subject_edu_semestr_id','edu_semestr_subject');
-        $this->dropForeignKey('se_edu_semestr_subject_subject_id','edu_semestr_subject');
-        $this->dropForeignKey('te_edu_semestr_subject_subject_type_id','edu_semestr_subject');
+        $this->dropForeignKey('je_edu_semestr_subject_edu_semestr_id', 'edu_semestr_subject');
+        $this->dropForeignKey('se_edu_semestr_subject_subject_id', 'edu_semestr_subject');
+        $this->dropForeignKey('te_edu_semestr_subject_subject_type_id', 'edu_semestr_subject');
         $this->dropTable('edu_semestr_subject');
     }
